+++ crates/plugin_minify/src/minify_resource_pot.rs
@@ -68,6 +68,7 @@ pub fn minify_js(
         &ExtraOptions {
           unresolved_mark,
           top_level_mark,
+          mangle_name_cache: None,
         },
       )
       .expect_module()