+++ crates/plugin_partial_bundling/src/merge_module_pots/test_boundaries.rs
@@ -1,5 +1,8 @@
-use farmfe_core::config::partial_bundling::PartialBundlingConfig;
-use farmfe_testing_helpers::construct_test_module_graph_complex;
+use farmfe_core::{
+  config::{partial_bundling::PartialBundlingConfig, Config},
+  resource::resource_pot::ResourcePot,
+};
+use farmfe_testing_helpers::{assert_resource_pots, construct_test_module_graph_complex};
 
 use super::{common::create_test_module_pot, merge_module_pots, ModuleGroupModulePots};
 
@@ -9,10 +12,10 @@ fn test_boundaries_enforce_min_size_not_enough_size() {
   let mut module_group_module_pots = ModuleGroupModulePots::new("B".into());
 
   let size = 10 * 1024;
-  let module_bucket_b = create_test_module_pot(&mut module_graph, &"B".into(), size, false);
-  let module_bucket_e = create_test_module_pot(&mut module_graph, &"E".into(), size, false);
-  let module_bucket_d = create_test_module_pot(&mut module_graph, &"D".into(), size * 2, true);
-  let module_bucket_h = create_test_module_pot(&mut module_graph, &"H".into(), size, true);
+  let module_bucket_b = create_test_module_pot(&mut module_graph, &"B".into(), "B", size, false);
+  let module_bucket_e = create_test_module_pot(&mut module_graph, &"E".into(), "E", size, false);
+  let module_bucket_d = create_test_module_pot(&mut module_graph, &"D".into(), "D", size * 2, true);
+  let module_bucket_h = create_test_module_pot(&mut module_graph, &"H".into(), "H", size, true);
 
   module_group_module_pots
     .add_module_pots("B_E".to_string(), vec![module_bucket_b, module_bucket_e]);
@@ -21,10 +24,13 @@ fn test_boundaries_enforce_min_size_not_enough_size() {
 
   let mut resource_pots = merge_module_pots(
     module_group_module_pots,
-    &PartialBundlingConfig {
-      target_concurrent_requests: 2,
-      target_min_size: 30 * 1024,
-      enforce_target_min_size: true,
+    &Config {
+      partial_bundling: Box::new(PartialBundlingConfig {
+        target_concurrent_requests: 2,
+        target_min_size: 30 * 1024,
+        enforce_target_min_size: true,
+        ..Default::default()
+      }),
       ..Default::default()
     },
     "B",
@@ -34,10 +40,8 @@ fn test_boundaries_enforce_min_size_not_enough_size() {
   resource_pots.sort_by_key(|p| p.id.clone());
 
   assert_eq!(resource_pots.len(), 2);
-  assert_eq!(resource_pots[0].immutable, true);
-  assert_eq!(resource_pots[0].modules(), vec![&"D".into(), &"H".into()]);
-  assert_eq!(resource_pots[1].immutable, false);
-  assert_eq!(resource_pots[1].modules(), vec![&"B".into(), &"E".into()]);
+
+  assert_resource_pots!(&resource_pots);
 }
 
 #[test]
@@ -46,10 +50,10 @@ fn test_boundaries_enforce_min_size_0() {
   let mut module_group_module_pots = ModuleGroupModulePots::new("B".into());
 
   let size = 10 * 1024;
-  let module_bucket_b = create_test_module_pot(&mut module_graph, &"B".into(), size, false);
-  let module_bucket_e = create_test_module_pot(&mut module_graph, &"E".into(), size, false);
-  let module_bucket_d = create_test_module_pot(&mut module_graph, &"D".into(), size * 2, true);
-  let module_bucket_h = create_test_module_pot(&mut module_graph, &"H".into(), size, true);
+  let module_bucket_b = create_test_module_pot(&mut module_graph, &"B".into(), "B", size, false);
+  let module_bucket_e = create_test_module_pot(&mut module_graph, &"E".into(), "E", size, false);
+  let module_bucket_d = create_test_module_pot(&mut module_graph, &"D".into(), "D", size * 2, true);
+  let module_bucket_h = create_test_module_pot(&mut module_graph, &"H".into(), "H", size, true);
 
   module_group_module_pots
     .add_module_pots("B_E".to_string(), vec![module_bucket_b, module_bucket_e]);
@@ -59,9 +63,12 @@ fn test_boundaries_enforce_min_size_0() {
   // test requests = 3
   let mut resource_pots = merge_module_pots(
     module_group_module_pots.clone(),
-    &PartialBundlingConfig {
-      target_concurrent_requests: 3,
-      target_min_size: 0,
+    &Config {
+      partial_bundling: Box::new(PartialBundlingConfig {
+        target_concurrent_requests: 3,
+        target_min_size: 0,
+        ..Default::default()
+      }),
       ..Default::default()
     },
     "B",
@@ -70,21 +77,18 @@ fn test_boundaries_enforce_min_size_0() {
 
   resource_pots.sort_by_key(|p| p.id.clone());
 
-  assert_eq!(resource_pots.len(), 3);
-  assert_eq!(resource_pots[0].immutable, false);
-  assert_eq!(resource_pots[0].modules(), vec![&"B".into(), &"E".into()]);
-  assert_eq!(resource_pots[1].immutable, true);
-  assert_eq!(resource_pots[1].modules(), vec![&"D".into()]);
-  assert_eq!(resource_pots[2].immutable, true);
-  assert_eq!(resource_pots[2].modules(), vec![&"H".into()]);
+  assert_resource_pots!(&resource_pots);
 
   // test requests = 4 and immutable_modules_weight = 0.5
   let mut resource_pots = merge_module_pots(
     module_group_module_pots.clone(),
-    &PartialBundlingConfig {
-      target_concurrent_requests: 4,
-      target_min_size: 0,
-      immutable_modules_weight: 0.5,
+    &Config {
+      partial_bundling: Box::new(PartialBundlingConfig {
+        target_concurrent_requests: 4,
+        target_min_size: 0,
+        immutable_modules_weight: 0.5,
+        ..Default::default()
+      }),
       ..Default::default()
     },
     "B",
@@ -94,14 +98,7 @@ fn test_boundaries_enforce_min_size_0() {
   resource_pots.sort_by_key(|p| p.id.clone());
 
   assert_eq!(resource_pots.len(), 4);
-  assert_eq!(resource_pots[0].immutable, true);
-  assert_eq!(resource_pots[0].modules(), vec![&"D".into()]);
-  assert_eq!(resource_pots[1].immutable, true);
-  assert_eq!(resource_pots[1].modules(), vec![&"H".into()]);
-  assert_eq!(resource_pots[2].immutable, false);
-  assert_eq!(resource_pots[2].modules(), vec![&"E".into()]);
-  assert_eq!(resource_pots[3].immutable, false);
-  assert_eq!(resource_pots[3].modules(), vec![&"B".into()]);
+  assert_resource_pots!(&resource_pots);
 }
 
 #[test]
@@ -110,10 +107,10 @@ fn test_boundaries_enforce_concurrent_requests_0() {
   let mut module_group_module_pots = ModuleGroupModulePots::new("B".into());
 
   let size = 10 * 1024;
-  let module_bucket_b = create_test_module_pot(&mut module_graph, &"B".into(), size, false);
-  let module_bucket_e = create_test_module_pot(&mut module_graph, &"E".into(), size, false);
-  let module_bucket_d = create_test_module_pot(&mut module_graph, &"D".into(), size * 2, true);
-  let module_bucket_h = create_test_module_pot(&mut module_graph, &"H".into(), size, true);
+  let module_bucket_b = create_test_module_pot(&mut module_graph, &"B".into(), "B", size, false);
+  let module_bucket_e = create_test_module_pot(&mut module_graph, &"E".into(), "E", size, false);
+  let module_bucket_d = create_test_module_pot(&mut module_graph, &"D".into(), "D", size * 2, true);
+  let module_bucket_h = create_test_module_pot(&mut module_graph, &"H".into(), "H", size, true);
 
   module_group_module_pots
     .add_module_pots("B_E".to_string(), vec![module_bucket_b, module_bucket_e]);
@@ -122,9 +119,12 @@ fn test_boundaries_enforce_concurrent_requests_0() {
 
   let mut resource_pots = merge_module_pots(
     module_group_module_pots.clone(),
-    &PartialBundlingConfig {
-      target_concurrent_requests: 0,
-      target_min_size: 10 * 1024,
+    &Config {
+      partial_bundling: Box::new(PartialBundlingConfig {
+        target_concurrent_requests: 0,
+        target_min_size: 10 * 1024,
+        ..Default::default()
+      }),
       ..Default::default()
     },
     "B",
@@ -134,20 +134,17 @@ fn test_boundaries_enforce_concurrent_requests_0() {
   resource_pots.sort_by_key(|p| p.id.clone());
 
   assert_eq!(resource_pots.len(), 4);
-  assert_eq!(resource_pots[0].immutable, true);
-  assert_eq!(resource_pots[0].modules(), vec![&"D".into()]);
-  assert_eq!(resource_pots[1].immutable, true);
-  assert_eq!(resource_pots[1].modules(), vec![&"H".into()]);
-  assert_eq!(resource_pots[2].immutable, false);
-  assert_eq!(resource_pots[2].modules(), vec![&"E".into()]);
-  assert_eq!(resource_pots[3].immutable, false);
-  assert_eq!(resource_pots[3].modules(), vec![&"B".into()]);
+
+  assert_resource_pots!(&resource_pots);
 
   let mut resource_pots = merge_module_pots(
     module_group_module_pots.clone(),
-    &PartialBundlingConfig {
-      target_concurrent_requests: 1,
-      target_min_size: 10 * 1024,
+    &Config {
+      partial_bundling: Box::new(PartialBundlingConfig {
+        target_concurrent_requests: 1,
+        target_min_size: 10 * 1024,
+        ..Default::default()
+      }),
       ..Default::default()
     },
     "B",
@@ -157,10 +154,5 @@ fn test_boundaries_enforce_concurrent_requests_0() {
   resource_pots.sort_by_key(|p| p.id.clone());
 
   assert_eq!(resource_pots.len(), 3);
-  assert_eq!(resource_pots[0].immutable, false);
-  assert_eq!(resource_pots[0].modules(), vec![&"B".into(), &"E".into()]);
-  assert_eq!(resource_pots[1].immutable, true);
-  assert_eq!(resource_pots[1].modules(), vec![&"D".into()]);
-  assert_eq!(resource_pots[2].immutable, true);
-  assert_eq!(resource_pots[2].modules(), vec![&"H".into()]);
+  assert_resource_pots!(&resource_pots);
 }