+++ packages/core/src/utils/cacheDir.ts
@@ -1,0 +1,11 @@
+import fs from 'node:fs';
+
+export async function isCacheDirExists(dir: string): Promise<boolean> {
+  try {
+    const hasCacheDir = fs.readdirSync(dir, { withFileTypes: true });
+
+    return !!(hasCacheDir && hasCacheDir.length);
+  } catch (_) {
+    return false;
+  }
+}